pyproject.toml
[project]
name = "open-duelyst-assets"
version = "0.1.0"
description = "Spritesheets and frames from OpenDuelyst"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.0.0",
]
